feat: BGMManager
인풋에 따라 스테이지 배경음악을 다르게 하는 코드입니다.
