+++ Gravity Controller/Assets/Scripts/UI/BGMManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9338ae2aab3d5ba4e8dd34903a3a7b2d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
